+++ conanfile.py
@@ -1,9 +1,11 @@
 from conans import ConanFile, CMake
+from conans.tools import load
+import re
+import os
 
 
 class bertrandConan(ConanFile):
     name = "bertrand"
-    version = "0.0.11"
     license = "LGPLv3"
     author = "Dominik Berner <dominik.berner+bertrand-conan@gmail.com"
     url = "https://github.com/bernedom/bertrand"
@@ -14,12 +16,21 @@ class bertrandConan(ConanFile):
     no_copy_source = True
     generators = "cmake", "txt", "cmake_find_package"
     build_requires = "Catch2/2.11.1@catchorg/stable"
+    _cmake = None
 
     def _configure_cmake(self):
-        cmake = CMake(self)
-        # Add additional settings with cmake.definitions["SOME_DEFINITION"] = True
-        cmake.configure()
-        return cmake
+        if self._cmake is None:
+            self._cmake = CMake(self)
+            # Add additional settings with cmake.definitions["SOME_DEFINITION"] = True
+            self._cmake.configure()
+        return self._cmake
+
+    def set_version(self):
+        cmake = load(os.path.join(self.recipe_folder, "CMakeLists.txt"))
+
+        version = re.search(
+            r"(?:[ \t]*)(?:VERSION\s+?)(\d+\.\d+\.\d+)", cmake).group(1)
+        self.version = version
 
     def build(self):
         cmake = self._configure_cmake()