+++ test/src/text_tests.cc
@@ -22,8 +22,8 @@ float divide(float dividend, float divisor) {
 TEST_CASE("GIVEN a contract with a message WHEN the contract fails THEN the "
           "message is passed on") {
   std::stringstream expected;
-  expected << __FILE__ << ":";
-  expected << __LINE__ + 1 << ": ('false') Cannot be false\n";
+  expected << "Assert at: " << __FILE__ << ":";
+  expected << __LINE__ + 1 << ": Require ('false') Cannot be false\n";
   REQUIRE_THROWS_WITH(Require(false, "Cannot be false"), expected.str());
 }
 
@@ -40,8 +40,8 @@ TEST_CASE("GIVEN an example function WHEN executed with invalid arguments THEN "
 TEST_CASE("GIVEN an example function WHEN executed with invalid arguments THEN "
           "contract is triggered AND message is set with correct location") {
   std::stringstream expected;
-  expected << __FILE__ << ":";
-  expected << 15 << ": ('divisor != 0') No division by zero\n";
+  expected << "Assert at: " << __FILE__ << ":";
+  expected << 15 << ": Require ('divisor != 0') No division by zero\n";
   REQUIRE_THROWS_WITH(divide(10.0f, 0.0f), expected.str());
 }
 
@@ -49,8 +49,8 @@ TEST_CASE("GIVEN a contract WHEN it fails THEN message is printed to stderr") {
   std::stringstream new_stderr;
   auto old_stderr = std::cerr.rdbuf(new_stderr.rdbuf());
   std::stringstream expected;
-  expected << __FILE__ << ":";
-  expected << __LINE__ + 1 << ": ('false') Cannot be false\n";
+  expected << "Assert at: " << __FILE__ << ":";
+  expected << __LINE__ + 1 << ": Require ('false') Cannot be false\n";
   REQUIRE_THROWS_WITH(Require(false, "Cannot be false"), expected.str());
   REQUIRE(expected.str().compare(new_stderr.str()) == 0);
   std::cerr.rdbuf(old_stderr);
@@ -61,8 +61,9 @@ TEST_CASE("GIVEN a contract AND contract has a variadic message WHEN it fails "
   std::stringstream new_stderr;
   auto old_stderr = std::cerr.rdbuf(new_stderr.rdbuf());
   std::stringstream expected;
-  expected << __FILE__ << ":";
-  expected << __LINE__ + 2 << ": ('false') Cannot be false with an argument\n";
+  expected << "Assert at: " << __FILE__ << ":";
+  expected << __LINE__ + 3
+           << ": Require ('false') Cannot be false with an argument\n";
   // clang-format off
   REQUIRE_THROWS_WITH(Require(false, "Cannot be false ", "with an argument"), expected.str());
   // clang-format on
@@ -75,9 +76,10 @@ TEST_CASE("GIVEN a contract AND contract has a variadic of multiple types WHEN "
   std::stringstream new_stderr;
   auto old_stderr = std::cerr.rdbuf(new_stderr.rdbuf());
   std::stringstream expected;
-  expected << __FILE__ << ":";
-  expected << __LINE__ + 3
-           << ": ('false') Cannot be false with an argument 99 123.11\n";
+  expected << "Assert at: " << __FILE__ << ":";
+  expected
+      << __LINE__ + 3
+      << ": Require ('false') Cannot be false with an argument 99 123.11\n";
   // clang-format off
   REQUIRE_THROWS_WITH(Require(false, "Cannot be false ", "with an argument ", 99, " ", 123.11), expected.str());
   // clang-format on